requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
distlib==0.3.9
dj-database-url==3.0.1
Django==5.2.1
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-extensions==4.1
django-js-asset==3.1.2
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
filelock==3.18.0
frozenlist==1.7.0
git-filter-repo==2.47.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
MarkupSafe==3.0.2
multidict==6.6.3
mysqlclient==2.2.7
numpy==2.2.6
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
razorpay==1.4.2
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==12.1.0
twilio==9.6.5
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
virtualenv==20.31.2
Werkzeug==3.1.3
whitenoise==6.9.0
yarl==1.20.1
zope.interface==7.2
gunicorn
whitenoise
dj-database-url
psycopg2-binary

